requirements.txt
openai
langchain
pinecone-client
python-dotenv
embeddings
playwright
streamlit
streamlit-chat
tiktoken
lark